// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  user: 'user',
  admin: 'admin',
  seller: 'seller',
  client: 'client',
  provider: 'provider'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ProductUnit = {
  liter: 'liter',
  cm: 'cm',
  gm: 'gm',
  kg: 'kg',
  piece: 'piece',
  seed: 'seed'
} as const

export type ProductUnit = (typeof ProductUnit)[keyof typeof ProductUnit]


export const ComponentUnit = {
  g_liter: 'g_liter',
  percentage: 'percentage',
  g_kg: 'g_kg',
  ppm: 'ppm',
  ppb: 'ppb'
} as const

export type ComponentUnit = (typeof ComponentUnit)[keyof typeof ComponentUnit]


export const Category = {
  herbicides: 'herbicides',
  insecticides: 'insecticides',
  fungicides: 'fungicides',
  acaricides: 'acaricides',
  nematicides: 'nematicides',
  growth_regulators: 'growth_regulators',
  fertilizers: 'fertilizers',
  seeds: 'seeds',
  tools: 'tools'
} as const

export type Category = (typeof Category)[keyof typeof Category]


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]